+++ CHANGELOG.md
@@ -115,7 +115,8 @@ GPU: 0
 
 ### Changed
 
--  **All `amd-smi set` options are now mutually exclusive**.  
+
+-  **All `amd-smi set` and `amd-smi reset` options are now mutually exclusive**.  
     - Users can only use one set option at a time now.  
 
 - **Python API for `amdsmi_get_energy_count()` will change the name for the `power` field to `energy_accumulator`**.